fix: simplify `_isActive` test and fix a bug in the test
I think this is a lot cleaner and fixes the issue where a `/section` link
is shown as active when `/other/section` is the actual url. This change
means all link paths passed MUST be absolute urls. However, I think
this is a good thing and the docs already reflect that.

cc @bebraw